Moved code to generate ticks and format function into d3-rendering/tick-marks as this could potentially be used for the DV graph. There is a full set off tests for it.
