+++ assets/src/scripts/utils.spec.js
@@ -1,6 +1,6 @@
 import {select} from 'd3-selection';
 import {
-    unicodeHtmlEntity, getHtmlFromString, deltaDays, replaceHtmlEntities, setEquality,
+    unicodeHtmlEntity, getHtmlFromString, replaceHtmlEntities, setEquality,
     wrap, mediaQuery, calcStartTime, callIf, parseRDB, convertFahrenheitToCelsius,
     convertCelsiusToFahrenheit, sortedParameters, getNearestTime} from './utils';
 
@@ -29,29 +29,6 @@ describe('Utils module', () => {
            expect(result).toContain('&#178;');
            expect(result).toContain('&#179;');
            expect(result.length).toBe(2);
-        });
-    });
-
-    describe('deltaDays', () => {
-
-        const date1 = new Date(2017, 2, 26, 16, 15, 0);
-        const date2 = new Date(2017, 3, 2, 4, 30, 0);
-        const date3 = new Date(2017, 3, 2, 21, 15, 0);
-        const date4 = new Date(2017, 2, 30, 16, 15, 0);
-
-        it('Returns the correct number of days when it is a bit less than an integer number of days', () => {
-           let result = deltaDays(date1, date2);
-           expect(result).toBe(7);
-        });
-
-        it('Returns the correct number of days when it is a bit more than an integer number of days', () => {
-           let result = deltaDays(date1, date3);
-           expect(result).toBe(7);
-        });
-
-        it('Returns the correct number of days when it is an exact integer number days', () => {
-           let result = deltaDays(date1, date4);
-           expect(result).toBe(4);
         });
     });
